Learned something about package manages
There is an ocaml package manager named `opam`. It is fairly easy to
use. But I had a hell of a time getting a little OpenGL snippet I stole
from Wikipedia to compile.

I first installed opam (it was in pacman repository, so no issues here).

Running `opam` with no arguments gives a nice list of options. I ran
`opam init`. I chose, foolishly, not to let it automatically configure
my files.

Then I installed the required OpenGL binder, `lablgl`. I added in a
bunch of flags for the compiler, linking the libraries. Then tried to
compile it. It died. Saying it had no idea where the thing I just
installed was.

Finally I tracked the files down and had to hand link them. Irritating,
but probably the sort of problem I won't have later.

The code for running the animation is pretty ugly. But I suppose it is
more succinct than similar programs in other languages. I'll look over
the details in the future, this was mostly a test of compiling.